Make skiboot able to be 2MB rather than 1MB
When built with gcov, skiboot is >1MB (closer to 1.5MB) and there
were a few assumptions about skiboot being <1MB.

The biggest one was that when code got larger, we'd have the sbss section
start in the middle of code, so that when we were going to relocate ourselves,
we'd only get the first 1MB of skiboot relocated, which excluded the
_DYNAMIC section, meaning that relocate() would not find the right
sections.

We also needed to not start writing over random parts of skiboot.

Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>