hw/p8-i2c: Fix OCC locking
There's a few issues with the Host<->OCC I2C bus handshaking. First up,
skiboot is currently examining the wrong bit when checking if the OCC
is currently using the bus. Secondly, when we need to wait for the OCC
to release the bus we are scheduling a recovery timer to run zero
timebase ticks after the current moment so the next poller run will
always time out the current request. The is also an issue with the
order of operations in the recovery timeout handler which can cause
an assertion failure when the cause of the recovery is a timeout
waiting for the OCC.

This patch addresses all these issues and sets the recovery timeout to
10ms.

Signed-off-by: Oliver O'Halloran <oohall@gmail.com>
[stewart@linux.vnet.ibm.com: PR_DEBUG rather than PR_ERR when OCC has lock]
Signed-off-by: Stewart Smith <stewart@linux.vnet.ibm.com>