firmware-versions: Add test case for parsing VERSION
Also make it possible to use with afl-lop/afl-fuzz just to help make
*sure* we're all good.

Additionally, if we hit a entry in VERSION that is larger than our
buffer size, we skip over it gracefully rather than overwriting the
stack. This is only a problem if VERSION isn't trusted, which as of
4b8cc05a94513816d43fb8bd6178896b430af08f it is verified as part of
Secure Boot.

CC: stable # v5.9+
Fixes: 9727fe384b8685270d344201f7e051475eea3a0b
Signed-off-by: Stewart Smith <stewart@linux.ibm.com>