core/cpu.c: avoid container_of(NULL) in next_cpu()
A certain finicky static analysis tool did point out that we were
operating on a value that could be null (and since first_cpu() calls
next_cpu(NULL) to get the first one, it also gets to be complained about
as next_cpu() could act on that NULL pointer).

So, rework things to shut the static analysis tool up, when in fact this
was never a problem.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>