Change ifdef around dump_fdt() to shut up static analysis
This is a dumb warning from a certain static analysis tool that a
function has no effect when the ifdef that would make it have an effect
isn't defined and we replace it with a no-op impl.

Putting the #ifdef around the call just so I don't have to discount this
damn static analysis false positive every time I go and look at the
results.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>