core/vm: try to handle recoverable MCEs by turning off VMM
This is not foolproof because local mappings and io accessors will
not cope with the VMM being switched off, but it may catch some
cases. io accessors in particular won't recover properly because
they won't switch to use the real-mode cache inhibited instructions
after VMM turns off. That could be solved by marking IO accessors
as non-recoverable, or providing an exception recovery address
which retries.

This is mostly just a debugging aid, real MCE recovery for VMM
faults would have to flush SLB and translation caches.

Signed-off-by: Stewart Smith <stewart@linux.ibm.com>